Refactorings split in multiple files
