+++ client/src/app/refactoring/refactoring-commands/add-message-broker.ts
@@ -1,0 +1,53 @@
+import { Graph } from "src/app/graph/model/graph";
+import { Refactoring } from "./refactoring-command";
+import { SmellObject } from "../smell";
+import * as joint from "jointjs";
+
+export class AddMessageBrokerRefactoring implements Refactoring {
+    links: joint.shapes.microtosca.RunTimeLink[];
+    graph: Graph;
+    team: joint.shapes.microtosca.SquadGroup;
+
+    addedSourceTargetbrokers: [joint.shapes.microtosca.Node, joint.shapes.microtosca.Node, joint.shapes.microtosca.CommunicationPattern][];
+
+    constructor(graph: Graph, smell: SmellObject) {
+        // For wobbly service interaction, adding message broker is disable whe the target node is a communication pattern
+        this.links = smell.getLinkBasedCauses().filter((link) => !(link.getTargetElement() instanceof joint.shapes.microtosca.CommunicationPattern))
+
+        this.graph = graph;
+        this.addedSourceTargetbrokers = [];
+    }
+
+    execute() {
+        let len = this.links.length;
+        for (var _i = 0; _i < len; _i++) {
+            var link = this.links.pop();
+            let sourceNode = <joint.shapes.microtosca.Node>link.getSourceElement();
+            let targetNode = <joint.shapes.microtosca.Node>link.getTargetElement();
+            let messageRouter = this.graph.addMessageBroker(`${sourceNode.getName()} ${targetNode.getName()}`);
+            this.graph.addRunTimeInteraction(sourceNode, messageRouter);
+            this.graph.addRunTimeInteraction(targetNode, messageRouter);
+            this.addedSourceTargetbrokers.push([sourceNode, targetNode, messageRouter]);
+            link.remove();
+        }
+    }
+
+    unexecute() {
+        let len = this.addedSourceTargetbrokers.length;
+        for (var _i = 0; _i < len; _i++) {
+            let sourceTargetPattern = this.addedSourceTargetbrokers.pop();
+            let link = this.graph.addRunTimeInteraction(sourceTargetPattern[0], sourceTargetPattern[1]);
+            this.links.push(link);
+            sourceTargetPattern[2].remove();
+        };
+    }
+
+    getName() {
+        return "Add message broker";
+    }
+
+    getDescription() {
+        return "Add message broker between two services";
+    }
+
+}
