+++ client/src/app/refactoring/refactoring-commands/add-service-discovery.ts
@@ -1,0 +1,32 @@
+import { Graph } from "src/app/graph/model/graph";
+import { Refactoring } from "./refactoring-command";
+import { SmellObject } from "../smell";
+
+export class AddServiceDiscoveryRefactoring implements Refactoring {
+    links: joint.shapes.microtosca.RunTimeLink[];
+
+    constructor(graph: Graph, smell: SmellObject) {
+        this.links = smell.getLinkBasedCauses();
+    }
+
+    execute() {
+        this.links.forEach(link => {
+            link.setDynamicDiscovery(true);
+        })
+    }
+
+    unexecute() {
+        this.links.forEach(link => {
+            link.setDynamicDiscovery(false);
+        })
+    }
+
+    getName() {
+        return "Add service discovery";
+    }
+
+    getDescription() {
+        return "Add service discovery";
+    }
+
+}
