+++ client/src/app/refactoring/refactoring-commands/merge-services-refactoring.ts
@@ -1,0 +1,94 @@
+import { Graph } from "src/app/graph/model/graph";
+import { Refactoring } from "./refactoring-command";
+import { SmellObject } from "../smell";
+
+export class MergeServicesRefactoring implements Refactoring {
+
+    smell: SmellObject;
+    graph: Graph;
+    sharedDatastore: joint.shapes.microtosca.Datastore;
+    mergedService: joint.shapes.microtosca.Service;
+    team: joint.shapes.microtosca.SquadGroup;
+
+    serviceIngoingOutgoing: [joint.shapes.microtosca.Node, joint.shapes.microtosca.Node[], joint.shapes.microtosca.Node[]][];
+
+
+    constructor(graph: Graph, smell: SmellObject) {
+        this.smell = smell;
+        this.graph = graph;
+        this.sharedDatastore = <joint.shapes.microtosca.Datastore>smell.getNodeBasedCauses()[0];
+        this.serviceIngoingOutgoing = [];
+        this._saveIncomingOutcomingNodes();
+
+    }
+
+    execute() {
+        this.mergedService = this.graph.addService("Merged Service");
+        this.graph.addRunTimeInteraction(this.mergedService, this.sharedDatastore);
+        this._addLinkToMergedService();
+        this._removeServicesAccessingDB();
+    }
+
+    unexecute() {
+        this._restoreServicesAccesingDB();
+        this._restoreLinks();
+        this.mergedService.remove();
+    }
+
+    getName() {
+        return "Merge services";
+    }
+
+    getDescription() {
+        return "Merge the services accessing the same database";
+    }
+
+    _restoreServicesAccesingDB() {
+        this.serviceIngoingOutgoing.forEach(nodeingoingOutgoing => {
+            let nameDeletedService = nodeingoingOutgoing[0].getName();
+            let service = this.graph.addService(nameDeletedService);
+            this.graph.addRunTimeInteraction(service, this.sharedDatastore);
+        });
+    }
+
+    _restoreLinks() {
+        this.serviceIngoingOutgoing.forEach(nodeingoingOutgoing => {
+            let nameDeletedService = nodeingoingOutgoing[0].getName();
+            let service = this.graph.findRootByName(nameDeletedService);
+            nodeingoingOutgoing[1].forEach(node => {
+                this.graph.addRunTimeInteraction(this.graph.findRootByName(node.getName()), service);
+            })
+            nodeingoingOutgoing[2].forEach(node => {
+                this.graph.addRunTimeInteraction(service, this.graph.findRootByName(node.getName()));
+            })
+        });
+    }
+
+    _removeServicesAccessingDB() {
+        this.serviceIngoingOutgoing.forEach(sio => {
+            sio[0].remove();
+        })
+    }
+
+    _addLinkToMergedService() {
+        this.serviceIngoingOutgoing.forEach(sio => {
+            // ingoing node
+            sio[1].forEach(node => {
+                this.graph.addRunTimeInteraction(node, this.mergedService)
+            })
+            // outgoing node
+            sio[2].forEach(node => {
+                this.graph.addRunTimeInteraction(this.mergedService, node)
+            })
+        })
+    }
+
+    _saveIncomingOutcomingNodes() {
+        this.smell.getLinkBasedCauses().forEach(link => {
+            let nodeAccessDB = <joint.shapes.microtosca.Node>link.getSourceElement();
+            let ingoing: joint.shapes.microtosca.Node[] = this.graph.getInboundNeighbors(nodeAccessDB);
+            let outgoing: joint.shapes.microtosca.Node[] = this.graph.getOutboundNeighbors(nodeAccessDB);
+            this.serviceIngoingOutgoing.push([nodeAccessDB, ingoing, outgoing]);
+        });
+    }
+}
