+++ client/src/app/refactoring/refactoring-commands/use-timeout.ts
@@ -1,0 +1,33 @@
+import { Graph } from "src/app/graph/model/graph";
+import { Refactoring } from "./refactoring-command";
+import { SmellObject } from "../smell";
+
+export class UseTimeoutRefactoring implements Refactoring {
+    links: joint.shapes.microtosca.RunTimeLink[];
+    graph: Graph;
+
+    constructor(graph: Graph, smell: SmellObject) {
+        this.links = smell.getLinkBasedCauses();
+        this.graph = graph;
+    }
+
+    execute() {
+        this.links.forEach(link => {
+            link.setTimedout(true);
+        });
+    }
+
+    unexecute() {
+        this.links.forEach(link => {
+            link.setTimedout(false);
+        });
+    }
+
+    getName() {
+        return "Use timeout";
+    }
+
+    getDescription() {
+        return "Use timeout";
+    }
+}
